Fix rendering issues with post meta-data & lists (ol, ul)
The meta-data for posts was being forced above the title for the post,
and forced to all upper-case.  We now no longer absolutely position
the meta-data, nor force the text to all upper-case.

Additionally, lists were rendering with the bullets to the left of the
left margin of the paragraph, with the start of the text matching the
left margin of paragraphs around it.  By no longer setting the left
margin explicitly to 0, the lists will start with the decoration
lining up with the left margin of the paragraphs around them, and the
text "indented".