Add a 301 redirect for PDX Hackathon's Google Group
This provides a much easier URL to send people to in order to get them
registered for the mailing list, as well as a permanent location that
can be pointed elsewhere.

Also, Freenode's group registration forms didn't like providing the
google group URL as the URL for the group to register.